@@ -3,36 +3,38 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "next dev",
+    "dev": "next dev --turbo",
     "build": "next build",
     "start": "next start",
     "lint": "next lint"
   },
   "dependencies": {
-    "@nextui-org/react": "^2.1.13",
-    "@react-email/button": "0.0.10",
-    "@react-email/components": "^0.0.7",
-    "@react-email/html": "0.0.4",
-    "@react-email/tailwind": "0.0.9",
-    "@types/node": "20.5.0",
-    "@types/react": "18.2.20",
-    "@types/react-dom": "18.2.7",
-    "autoprefixer": "10.4.15",
-    "encoding": "^0.1.13",
-    "eslint": "8.47.0",
-    "eslint-config-next": "13.4.16",
+    "@nextui-org/react": "^2.2.8",
+    "@react-email/button": "0.0.11",
+    "@react-email/components": "^0.0.11",
+    "@react-email/html": "0.0.6",
+    "@react-email/tailwind": "0.0.12",
     "framer-motion": "^10.16.4",
-    "next": "^13.5.4",
+    "next": "^14.0.2",
     "next-themes": "^0.2.1",
-    "postcss": "8.4.31",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-email": "1.9.5",
     "react-hot-toast": "^2.4.1",
     "react-icons": "^4.11.0",
     "react-type-animation": "^3.2.0",
-    "resend": "^1.1.0",
-    "tailwindcss": "3.3.3",
+    "resend": "^2.0.0"
+  },
+  "devDependencies": {
+    "encoding": "^0.1.13",
+    "postcss": "8.4.31",
+    "eslint": "8.53.0",
+    "eslint-config-next": "14.0.2",
+    "@types/react": "18.2.37",
+    "@types/react-dom": "18.2.15",
+    "autoprefixer": "10.4.16",
+    "tailwindcss": "3.3.5",
+    "@types/node": "20.9.0",
     "typescript": "5.2.2"
   }
 }
