@@ -1,8 +1,8 @@
 export const NAV_LINKS = [
   { title: "Home", href: "/" },
-  { title: "About", href: "#about" },
-  { title: "Projects", href: "#projects" },
-  { title: "Contact", href: "#contact" },
+  { title: "About", href: "/about" },
+  { title: "Projects", href: "/projects" },
+  { title: "Contact", href: "/contact" },
 ] as const;
 
 export const PROJECTS = [
